requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# machine learning
tensorflow-gpu==2.1.0
Pillow==7.0.0
numpy==1.18.1
scikit-learn==0.22.1
pandas==1.0.1
seaborn==0.10.0
Flask==1.1.1
